+++ src/Poseidon/CLI/Genoconvert.hs
@@ -43,7 +43,7 @@ data GenoconvertOptions = GenoconvertOptions
 
 pacReadOpts :: PackageReadOptions
 pacReadOpts = defaultPackageReadOptions {
-        _readOptStopOnDuplicates = False
+          _readOptStopOnDuplicates = False
         , _readOptIgnoreChecksums  = True
         , _readOptIgnoreGeno       = False
         , _readOptGenoCheck        = True